Extract maintenance SQL and split index modules
