+++ src/sql/maintenance/clustered_incremental_index_setup.sql
@@ -1,0 +1,5 @@
+CREATE TABLE {{fts_schema}}.{{terms_storage_table}} AS
+SELECT termid, docid, fieldid FROM {{fts_schema}}.terms;
+DROP TABLE {{fts_schema}}.terms;
+CREATE VIEW {{fts_schema}}.terms AS
+SELECT termid, docid, fieldid FROM {{fts_schema}}.{{terms_storage_table}};
