+++ src/sql/maintenance/clustered_insert_token_ctes.sql
@@ -1,0 +1,13 @@
+WITH {{new_docs_cte}},
+tokenized AS (
+    {{union_fields_query}}
+),
+stemmed_stopped AS (
+    SELECT stem(t.w, {{stemmer}}) AS term,
+           t.docid AS docid,
+           t.fieldid AS fieldid
+    FROM tokenized AS t
+    WHERE t.w NOT NULL
+      AND t.w <> ''
+      AND t.w NOT IN (SELECT sw FROM {{fts_schema}}.stopwords)
+)
