+++ src/sql/maintenance/incremental_index_setup.sql
@@ -1,0 +1,1 @@
+CREATE UNIQUE INDEX {{index_name}} ON {{fts_schema}}.docs(name);
