+++ src/sql/maintenance/layered_affected_term_ids.sql
@@ -1,0 +1,3 @@
+SELECT DISTINCT d.termid
+FROM stemmed_stopped AS ss
+JOIN {{fts_schema}}.dict AS d ON ss.term = d.term
