+++ src/sql/maintenance/tokenize_deleted_field.sql
@@ -1,0 +1,2 @@
+SELECT unnest({{fts_schema}}.tokenize(fts_di.{{input_value}})) AS w
+FROM fts_old_rows AS fts_di
